[NorePanel] Added Flow in/out pin type

+++ Stimpi/Stimpi/src/Stimpi/Graphics/BufferObject.h
@@ -34,13 +34,13 @@ namespace Stimpi
 		VertexBuffer(uint32_t capacity, uint32_t vertexSize) : m_Data({}), m_Capacity(capacity), m_VertexSize(vertexSize) {}
 
 		float* GetRawData() { return m_Data.data(); }
-		uint32_t GetCount() { return m_Data.size() / m_VertexSize; }
+		size_t GetCount() { return m_Data.size() / m_VertexSize; }
 
 		std::vector<float>& GetData() { return m_Data; }
 
 	private:
 		std::vector<float> m_Data;
-		uint32_t m_Capacity;
-		uint32_t m_VertexSize;
+		size_t m_Capacity;
+		size_t m_VertexSize;
 	};
 }
